Clean up Particle System on Car Controller, FSM for Car Controller In but bug (report in descr)
The car controller has a FSM system ( to emit particles and later to help with car sound), which works on both acceleration values and steering. Acceleration and steering are great, however distinguishing between braking and reversing is bugged due to not getting the forward vector properly. Will discuss fix in next team meeting.